+++ test/pvp-arena.js
@@ -1,15 +1,15 @@
-const web3 = require('web3');
-const { expectRevert } = require('@openzeppelin/test-helpers');
-const helpers = require('./helpers');
+const web3 = require("web3");
+const { expectRevert, time } = require("@openzeppelin/test-helpers");
+const helpers = require("./helpers");
 
-const SkillToken = artifacts.require('SkillToken');
-const Characters = artifacts.require('Characters');
-const Weapons = artifacts.require('Weapons');
-const Shields = artifacts.require('Shields');
-const PvpArena = artifacts.require('PvpArena');
-const BasicPriceOracle = artifacts.require('BasicPriceOracle');
+const SkillToken = artifacts.require("SkillToken");
+const Characters = artifacts.require("Characters");
+const Weapons = artifacts.require("Weapons");
+const Shields = artifacts.require("Shields");
+const PvpArena = artifacts.require("PvpArena");
+const BasicPriceOracle = artifacts.require("BasicPriceOracle");
 
-contract('PvpArena', accounts => {
+contract("PvpArena", (accounts) => {
   let pvpArena, characters, weapons;
   beforeEach(async () => {
     skillToken = await SkillToken.deployed();
@@ -25,116 +25,205 @@ contract('PvpArena', accounts => {
     await weapons.grantRole(await weapons.GAME_ADMIN(), accounts[0]);
     await shields.grantRole(await shields.GAME_ADMIN(), accounts[0]);
 
-    await skillToken.transferFrom(skillToken.address, accounts[1], web3.utils.toWei('10', 'ether'));
-    
-    await priceOracle.setCurrentPrice('10');
+    await skillToken.transferFrom(
+      skillToken.address,
+      accounts[1],
+      web3.utils.toWei("10", "ether")
+    );
+
+    await priceOracle.setCurrentPrice("10");
   });
 
-
-  describe('#getArenaTier', () => {
+  describe("#getArenaTier", () => {
     let character0Id;
     let character1Id;
-    
+
     beforeEach(async () => {
-      character0Id = await helpers.createCharacter(accounts[1], '123', { characters });
-      character1Id = await helpers.createCharacter(accounts[1], '456', { characters });
-      
+      character0Id = await helpers.createCharacter(accounts[1], "123", {
+        characters,
+      });
+      character1Id = await helpers.createCharacter(accounts[1], "456", {
+        characters,
+      });
+
       await characters.gainXp(character0Id, 190);
       await characters.gainXp(character1Id, 220);
     });
 
-    it('should return the right arena tier', async () => {
-      const character0arenaTier = await pvpArena.getArenaTier(character0Id, { from: accounts[1] });
-      const character1arenaTier = await pvpArena.getArenaTier(character1Id, { from: accounts[1] });
-      expect(character0arenaTier.toString()).to.equal('0');
-      expect(character1arenaTier.toString()).to.equal('1');
+    it("should return the right arena tier", async () => {
+      const character0arenaTier = await pvpArena.getArenaTier(character0Id, {
+        from: accounts[1],
+      });
+      const character1arenaTier = await pvpArena.getArenaTier(character1Id, {
+        from: accounts[1],
+      });
+      expect(character0arenaTier.toString()).to.equal("0");
+      expect(character1arenaTier.toString()).to.equal("1");
     });
   });
 
-  describe('#enterArena', () => {
+  describe("#enterArena", () => {
     let characterId;
     let weaponId;
     let shieldId;
     let cost;
 
     beforeEach(async () => {
-      weaponId = await helpers.createWeapon(accounts[1], '123', { weapons });
-      shieldId = await helpers.createShield(accounts[1], '123', { shields });
-      characterId = await helpers.createCharacter(accounts[1], '123', { characters });
+      weaponId = await helpers.createWeapon(accounts[1], "123", { weapons });
+      shieldId = await helpers.createShield(accounts[1], "123", { shields });
+      characterId = await helpers.createCharacter(accounts[1], "123", {
+        characters,
+      });
     });
 
-    describe('happy path', async () => {
+    describe("happy path", async () => {
       let enterArenaReceipt;
 
       beforeEach(async () => {
-        cost = await pvpArena.getEntryCost(characterId, { from: accounts[1] });
-        await skillToken.approve(pvpArena.address, cost, { from: accounts[1] });
-        enterArenaReceipt = await pvpArena.enterArena(characterId, weaponId, { from: accounts[1] });
+        cost = await pvpArena.getEntryWager(characterId, { from: accounts[1] });
+        await skillToken.approve(pvpArena.address, web3.utils.toWei(cost), {
+          from: accounts[1],
+        });
+        enterArenaReceipt = await pvpArena.enterArena(
+          characterId,
+          weaponId,
+          0,
+          false,
+          {
+            from: accounts[1],
+          }
+        );
       });
 
-      it('should lock the entry cost');
+      it("should lock duelCost from wagered skill");
 
-      it('should add the character with its weapon and shield to the fighters list');
+      it("should add the character with its weapon and shield to the arena", async () => {
+        const isCharacterInArena = await pvpArena.isCharacterInArena(
+          characterId
+        );
+
+        expect(isCharacterInArena).to.equal(true);
+      });
     });
 
-    describe('character unavailable', async () => {
+    describe("character already in arena", async () => {
       let weapon2Id;
 
       beforeEach(async () => {
-        weaponId = await createWeapon();
-        weapon = await weapons.get(weaponId);
-        characterId = await createCharacter();
-        character = await characters.get(characterId);
+        await helpers.createWeapon(accounts[1], "123", { weapons });
+        weapon2Id = await helpers.createWeapon(accounts[1], "123", { weapons });
+        characterId = await helpers.createCharacter(accounts[1], "123", {
+          characters,
+        });
+        pvpArena.enterArena(characterId, weaponId, 0, false, {
+          from: accounts[1],
+        });
       });
 
-      it("shoul enter in arena:",async()=>{
-        const cost = await pvpArena.getEntryCost(characterId);
-        await skillToken.approve(pvpArena.address, cost,{from:accounts[1]});
-        enterArenaReceipt = await pvpArena.enterArena(characterId, weaponId,0,false,{from:accounts[1]});
-        let isCharacterUsed= await pvpArena.isCharacterUsed(characterId);
-        expect(isCharacterUsed).to.equal(true);
-      })
-
-      it('should revert', async () => {
-        await expectRevert(pvpArena.enterArena(characterId, weapon2Id, { from: accounts[1] }));
-      });
-    });
-
-    describe('weapon unavailable', async () => {
-      beforeEach(async () => {
-        character2Id = await helpers.createCharacter(accounts[1], '443', { characters });
-
-        cost = await pvpArena.getEntryCost(characterId, { from: accounts[1] });
-        await skillToken.approve(pvpArena.address, cost, { from: accounts[1] });
-        await pvpArena.enterArena(characterId, weaponId, { from: accounts[1] });
-      });
-
-      it('should revert', async () => {
-        await expectRevert(pvpArena.enterArena(character2Id, weaponId, { from: accounts[1] })); 
+      it("should revert", async () => {
+        await expectRevert(
+          pvpArena.enterArena(characterId, weapon2Id, 0, false, {
+            from: accounts[1],
+          }),
+          "The character is already in the arena"
+        );
       });
     });
 
-    describe('shield unavailable', () => {
+    describe("weapon already in arena", async () => {
       beforeEach(async () => {
-        character2Id = await helpers.createCharacter(accounts[1], '152', { characters });
-        weapon2Id = await helpers.createWeapon(accounts[1], '123', { weapons });
-        shieldId = await helpers.createShield(accounts[1], '446', { shields });
+        character2Id = await helpers.createCharacter(accounts[1], "443", {
+          characters,
+        });
 
-        cost = await pvpArena.getEntryCost(characterId, { from: accounts[1] });
-        await skillToken.approve(pvpArena.address, cost, { from: accounts[1] })
-        // await pvpArena.enterArena(characterId, weaponId, shieldId, { from: accounts[1] })
+        cost = await pvpArena.getEntryWager(characterId, { from: accounts[1] });
+
+        await skillToken.approve(pvpArena.address, web3.utils.toWei(cost), {
+          from: accounts[1],
+        });
+        await pvpArena.enterArena(characterId, weaponId, 0, false, {
+          from: accounts[1],
+        });
       });
 
-      it('should revert', async () => {
-        // await expectRevert(pvpArena.enterArena(character2Id, weapon2Id, shieldId, { from: accounts[1] }));
-      })
+      it("should revert", async () => {
+        cost = await pvpArena.getEntryWager(character2Id, {
+          from: accounts[1],
+        });
+        await skillToken.approve(pvpArena.address, web3.utils.toWei(cost), {
+          from: accounts[1],
+        });
+
+        await expectRevert(
+          pvpArena.enterArena(character2Id, weaponId, 0, false, {
+            from: accounts[1],
+          }),
+          "The weapon is already in the arena"
+        );
+      });
     });
 
-    describe('character is not sender\'s', () => {
+    describe("shield already in arena", () => {
       beforeEach(async () => {
-        cost = await pvpArena.getEntryCost(characterId, { from: accounts[1] });
-        await skillToken.approve(pvpArena.address, cost, { from: accounts[1] });
-        await pvpArena.enterArena(characterId, weaponId, { from: accounts[1] });
+        character2Id = await helpers.createCharacter(accounts[1], "152", {
+          characters,
+        });
+
+        weapon2Id = await helpers.createWeapon(accounts[1], "123", { weapons });
+        shieldId = await helpers.createShield(accounts[1], "446", { shields });
+
+        cost = await pvpArena.getEntryWager(character2Id, {
+          from: accounts[1],
+        });
+        await skillToken.approve(pvpArena.address, web3.utils.toWei(cost), {
+          from: accounts[1],
+        });
+        await pvpArena.enterArena(characterId, weaponId, shieldId, true, {
+          from: accounts[1],
+        });
+      });
+
+      it("should revert", async () => {
+        cost = await pvpArena.getEntryWager(character2Id, {
+          from: accounts[1],
+        });
+        await skillToken.approve(pvpArena.address, web3.utils.toWei(cost), {
+          from: accounts[1],
+        });
+
+        await expectRevert(
+          pvpArena.enterArena(character2Id, weapon2Id, shieldId, true, {
+            from: accounts[1],
+          }),
+          "The shield is already in the arena"
+        );
+      });
+    });
+
+    describe("character is not sender's", () => {
+      let otherCharacterId;
+
+      beforeEach(async () => {
+        otherCharacterId = await helpers.createCharacter(accounts[2], "123", {
+          characters,
+        });
+
+        cost = await pvpArena.getEntryWager(characterId, {
+          from: accounts[1],
+        });
+
+        await skillToken.approve(pvpArena.address, web3.utils.toWei(cost), {
+          from: accounts[1],
+        });
+      });
+
+      it("should revert", async () => {
+        await expectRevert(
+          pvpArena.enterArena(otherCharacterId, weaponId, 0, false, {
+            from: accounts[1],
+          }),
+          "You don't own the given character"
+        );
       });
     });
   });
